+++ src/fmmax/basis.py
@@ -72,6 +72,16 @@ class Expansion:
                 f"{self.basis_coefficients.shape}."
             )
 
+    def __hash__(self) -> str:
+        return hash(self.basis_coefficients.tobytes())
+
+    def __eq__(self, other: Any) -> bool:
+        if not isinstance(other, Expansion):
+            return False
+        if self.basis_coefficients.shape != other.basis_coefficients.shape:
+            return False
+        return onp.all(self.basis_coefficients == other.basis_coefficients)
+
     @property
     def num_terms(self) -> int:
         return self.basis_coefficients.shape[-2]